Move DerbySnapshotOperation and DistributedDatabaseOperateImpl to core module.
