+++ test/config-test/src/test/java/com/alibaba/nacos/test/base/BaseClusterTest.java
@@ -25,7 +25,7 @@ import com.alibaba.nacos.common.notify.Event;
 import com.alibaba.nacos.common.notify.NotifyCenter;
 import com.alibaba.nacos.common.notify.listener.Subscriber;
 import com.alibaba.nacos.persistence.model.event.RaftDbErrorEvent;
-import com.alibaba.nacos.config.server.service.repository.embedded.DistributedDatabaseOperateImpl;
+import com.alibaba.nacos.core.persistence.DistributedDatabaseOperateImpl;
 import com.alibaba.nacos.consistency.ProtocolMetaData;
 import com.alibaba.nacos.consistency.cp.CPProtocol;
 import com.alibaba.nacos.consistency.cp.MetadataKey;
